+++ lesson07/assignment/vue-cli/src/views/Trips.vue
@@ -20,43 +20,9 @@
 <script>
   import Card from '@/components/Card.vue'
   import Searchbar from '@/components/Searchbar.vue'
+  import { decks } from '../util.js'
 
-  //tried moving decks variable to main.js and importing into .vue
-  //import {decks} from 'main.js'
-  //get unused vars error - how do I fix this to only have this variable in one place?
 
-  var decks = [
-  {'id': 'belize',
-  'location': 'Belize',
-  'images':'belize.png',
-  'area': 'Central America',
-  'imagelink': '/travel-photos#belize', //this id section thing doesn't work correctly
-  text: "Picture of Belize"},
-  {'id': 'costarica',
-  'location': 'Costa Rica',
-  'images':'costarica.png',
-  'imagelink': '/travel-photos#costarica',
-  'area': 'Central America',
-  text: "Costa Rica - Manuel Antonio National Park"},
-  {'id': 'croatia',
-  'location': 'Croatia',
-  'images':'croatia.png',
-  'area': 'Europe',
-  'imagelink': '/travel-photos#croatia',
-  text: "Picture of Croatia"},
-  {'id': 'italy',
-  'location': 'Italy',
-  'images':'italy.png',
-  'area': 'Europe',
-  'imagelink': '/travel-photos#italy',
-  text: "Picture of Italy"},
-  {'id': 'thailand',
-  'location': 'Thailand',
-  'images':'thaibeach.png',
-  'area': 'Asia',
-  'imagelink': '/travel-photos#thailand',
-  text: "Picture of beach in Thailand"}
-  ];
   export default {
   name: 'Trips',
   components: {
@@ -65,7 +31,7 @@
   data(){
       return {
       decks: decks
-              }
+    }
   }
   };
